01. Preparing the dive
In this class, you learned how to:

Build a dynamic list with ListView.Builderin video 1.2: we saw that the Widget ListView is a mechanical way of building static lists and demands that we place items one by one. To optimize time, we use the ListView.builder, which creates a dynamic list according to the data received ( in the X file we leave the database with product information such as title, price, description etc );
Understand that the code may be right, but inadequate in relation to: although the IDE did not report any errors, we can improve the layout of the application so that it looks more like the Figma project.
We are starting the dive and we will see techniques for building several different types of lists.